apimarket specific func should be partitioned
