+++ src/index.js
@@ -1,13 +1,15 @@
 const Eos = require('eosjs');
 const accounts = require('./accounts');
 const cpu = require('./tokens/cpu');
+const crypto = require('./modules/crypto');
 const eos = require('./eos');
 const instrument = require('./instrument');
-const crypto = require('./modules/crypto');
+const offer = require('./apimarket/offer');
 const ore = require('./tokens/ore');
 const oreStandardToken = require('./orestandardtoken');
 const rightsRegistry = require('./rightsregistry');
-const usageLog = require('./usagelog');
+const voucher = require('./apimarket/voucher');
+const usageLog = require('./apimarket/usagelog');
 
 class Orejs {
   constructor(config = {}) {
@@ -19,10 +21,12 @@ class Orejs {
     Object.assign(this, crypto);
     Object.assign(this, eos);
     Object.assign(this, instrument);
+    Object.assign(this, offer);
     Object.assign(this, ore);
     Object.assign(this, oreStandardToken);
     Object.assign(this, rightsRegistry);
     Object.assign(this, usageLog);
+    Object.assign(this, voucher);
   }
 
   constructEos(config) {
